requirements.txt
pandas
numpy
google-play-scraper
chromadb
google-generativeai
sentence-transformers
torch
tqdm
